fix: extract component, and memo the logos so they dont reload all the time
